feat: add folder modules, add new mod
