+++ Food/js/modules/tabs.js
@@ -1,0 +1,42 @@
+function tabs() {
+  // Tabs
+  
+let tabs = document.querySelectorAll('.tabheader__item'),
+  tabsContent = document.querySelectorAll('.tabcontent'),
+  tabsParent = document.querySelector('.tabheader__items');
+
+function hideTabContent() {
+      
+      tabsContent.forEach(item => {
+          item.classList.add('hide');
+          item.classList.remove('show', 'fade');
+      });
+
+      tabs.forEach(item => {
+          item.classList.remove('tabheader__item_active');
+      });
+}
+
+function showTabContent(i = 0) {
+      tabsContent[i].classList.add('show', 'fade');
+      tabsContent[i].classList.remove('hide');
+      tabs[i].classList.add('tabheader__item_active');
+  }
+  
+  hideTabContent();
+  showTabContent();
+
+tabsParent.addEventListener('click', function(event) {
+  const target = event.target;
+  if(target && target.classList.contains('tabheader__item')) {
+          tabs.forEach((item, i) => {
+              if (target == item) {
+                  hideTabContent();
+                  showTabContent(i);
+              }
+          });
+  }
+  });
+}
+
+module.exports = tabs;
